moved voro++ files int src/ and renamed .hh -> .h
