+++ src/Makevars.R
@@ -1,0 +1,3 @@
+SOURCES = $(wildcard voropp046/*.cc)
+OBJECTS = libvoro++.o RcppExports.o $(SOURCES:.c=.o)
+
