+++ src/container_prd.cc
@@ -8,7 +8,7 @@
  * \brief Function implementations for the container_periodic_base and
  * related classes. */
 
-#include "container_prd.hh"
+#include "container_prd.h"
 
 namespace voro {
