+++ src/worklist.h
@@ -4,15 +4,15 @@
 // Email    : chr@alum.mit.edu
 // Date     : August 30th 2011
 
-/** \file worklist.hh
+/** \file worklist.h
  * \brief Header file for setting constants used in the block worklists that are
  * used during cell computation.
  *
  * This file is automatically generated by worklist_gen.pl and it is not
  * intended to be edited by hand. */
 
-#ifndef VOROPP_WORKLIST_HH
-#define VOROPP_WORKLIST_HH
+#ifndef VOROPP_WORKLIST_H
+#define VOROPP_WORKLIST_H
 
 namespace voro {
