+++ src/worklist_gen.pl
@@ -32,7 +32,7 @@ $d=2*$dis+1;$dd=$d*$d;$d0=(1+$d+$dd)*$dis;
 use Math::Trig;
 
 # Construct the worklist header file, based on the parameters above
-open W,">worklist.hh";
+open W,">worklist.h";
 print W <<EOF;
 // Voro++, a 3D cell-based Voronoi library
 //
@@ -40,15 +40,15 @@ print W <<EOF;
 // Email    : chr\@alum.mit.edu
 // Date     : August 30th 2011
 
-/** \\file worklist.hh
+/** \\file worklist.h
  * \\brief Header file for setting constants used in the block worklists that are
  * used during cell computation.
  *
  * This file is automatically generated by worklist_gen.pl and it is not
  * intended to be edited by hand. */
 
-#ifndef VOROPP_WORKLIST_HH
-#define VOROPP_WORKLIST_HH
+#ifndef VOROPP_WORKLIST_H
+#define VOROPP_WORKLIST_H
 
 namespace voro {
